// frontend/CRM/src/app/contacts/contacts.interface.ts
export interface Contact {
    id: number;
    createdTime: Date;
    firstName: string;
    lastName: string;
    birthday: string;
    age: number;
    email: string;
    phone:  string;
    state:  string;
    country:  string;
    city:  string;
    street:  string;
    houseNumber: string;
    zipCode: string;
    notes: string;
}
	